third_party/blink/web_tests/platform/generic/fast/html/meter-user-modify-expected.txt

PASS layoutTreeAfterType is layoutTreeBeforeType
PASS successfullyParsed is true

TEST COMPLETE


The inner element of meter should not be deleteable.

(before)In meter(after)
